python/requirements.txt
pyserial
customtkinter
mediapipe
opencv-contrib-python
pillow
